tweak: moving GameOptionsCard out of the ScrollView and into a Divider below; making it not a CollapsableCard
I think these are particularly important options that a player wants to be aware of pre any game creation. As the sidebar content grows, it became more easy for these options to be hidden. Argument for whether it should also be collapsable, but I didn't like the feel of one collapsable card scrolling under another. Atm I think these options are a good size to not hurt us too much on sidebar real estate.