feat: tweaking some styling which was causing cards to overflow the screen on my mobile
Gave some padding to the UpdateLog to have it be consistent with other cards on the smaller mobile screen. It looks like the issue was the cards were just taking a 400px width when that was more width than the screen had, so I've introduced a maxWidth of the screen size. Didn't have any effect elsewhere in the app