feat: 3d models for rookKnight and bishopKnight
sharing point descriptions of reused model parts. if there are special points on the part, the part is described as an object with features- so that the feature can be referenced and there is more clear coupling i.e. if we change the points we need to consider changing the features. several ways to go about these changes, I was also considering just having duplication of the parts- but thought sharing was more useful

I played around with the knight head: giving it a bump up part way through the nose length (inspired by a quick google search), rounding out under the end of the nose (seemed a bit blocky) and making the ear quite a bit shorter, the big ears stood out in a funny way on the knight heavy grand chess board