feat: split some findPaths functionality into findBasePaths
open to concept discussion and naming. The idea here is to have access to a basic/naive pathing which only cares about fundamentals like board geometry and piece blocking/capturing- it's almost an application of gaits to .path, just with the addition of handlePieceCollisions (which I'm deeming 'base', but maybe it can be in it's own layer).

This does add some extra structure (grouping of interruption points), but I think from a game design point of view we want something less granular than the interruption points- and cash in some genericness to have rules interact having awareness of groups rather than individual interruption points.