refactor: renaming keys to typecastKeys
I think we should be more explicit that the typecasting is happening. Object.keys is always returning a string[], which we are sometimes typecasting to a number[], symbol[]. In many cases this is helpful and non-eventful, but when there is an important mismatch it can be hard to spot due to the lack of appropriate type warnings.