better matching the 3d spherical board with flat board
having white at the bottom, and rotating the sphere so that the pieces line up. other variants are not lined up as nicely- it would be nice to do that at some stage, and also making the file sliding transformation stay the same between 2d and 3d boards. we also don't flip the 3d board for online play currently.