arrows and squares are looking good.
Undid the square arrow projection, realised it's important to have a pair of arrows back and forth look good, so adjusted params for that.

I think this look is nice. Something to note- when the square highlight is exactly the size of the square there appears to be a ~1 pixel difference on some squares depending on screen size- likely a rounding error of some kind- hard to pin down. It makes the overflow square highlight a clearer choice appearance wise. Also the current look is fine for overlaying a piece, it'll be a little more work to have the square highlight behind the piece, but arrow in front (split into seperate svgs, play with the zIndexing of the svg and pieces).